made a separate restcontroller
